Move VCM logging from VCM constructor to Configuration
