+++ kroxylicious-runtime/src/main/java/io/kroxylicious/proxy/config/Configuration.java
@@ -7,6 +7,7 @@ package io.kroxylicious.proxy.config;
 
 import java.util.ArrayList;
 import java.util.Collection;
+import java.util.Collections;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
@@ -16,6 +17,12 @@ import java.util.Optional;
 import java.util.Set;
 import java.util.function.Function;
 import java.util.stream.Collectors;
+
+import io.kroxylicious.proxy.config.tls.AllowDeny;
+import io.kroxylicious.proxy.config.tls.Tls;
+import io.kroxylicious.proxy.config.tls.TrustOptions;
+import io.kroxylicious.proxy.config.tls.TrustProvider;
+import io.kroxylicious.proxy.service.HostPort;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -166,13 +173,57 @@ public record Configuration(
                                                             @NonNull List<NamedFilterDefinition> filterDefinitions,
                                                             @NonNull String virtualClusterNodeName) {
 
-        return new VirtualClusterModel(virtualClusterNodeName,
+        VirtualClusterModel virtualClusterModel = new VirtualClusterModel(virtualClusterNodeName,
                 virtualCluster.targetCluster(),
                 buildAddressProviderService(virtualCluster, pfr),
                 virtualCluster.tls(),
                 virtualCluster.logNetwork(),
                 virtualCluster.logFrames(),
                 filterDefinitions);
+        logVirtualClusterSummary(virtualClusterModel.getClusterName(), virtualClusterModel.targetCluster(), virtualClusterModel.getClusterNetworkAddressConfigProvider(), virtualCluster.tls());
+        return virtualClusterModel;
+    }
+
+
+    @SuppressWarnings("java:S1874") // the classes are deprecated because we don't want them in the API module
+    private static void logVirtualClusterSummary(String clusterName,
+                                                 TargetCluster targetCluster,
+                                                 ClusterNetworkAddressConfigProvider clusterNetworkAddressConfigProvider,
+                                                 Optional<Tls> tls) {
+        try {
+            HostPort downstreamBootstrap = clusterNetworkAddressConfigProvider.getClusterBootstrapAddress();
+            var downstreamTlsSummary = generateTlsSummary(tls);
+
+            HostPort upstreamHostPort = targetCluster.bootstrapServersList().get(0);
+            var upstreamTlsSummary = generateTlsSummary(targetCluster.tls());
+
+            LOGGER.info("Virtual Cluster: {}, Downstream {}{} => Upstream {}{}",
+                    clusterName, downstreamBootstrap, downstreamTlsSummary, upstreamHostPort, upstreamTlsSummary);
+        }
+        catch (Exception e) {
+            LOGGER.warn("Failed to log summary for Virtual Cluster: {}", clusterName, e);
+        }
+    }
+
+    private static String generateTlsSummary(Optional<Tls> tlsToSummarize) {
+        var tls = tlsToSummarize.map(t -> Optional.ofNullable(t.trust())
+                        .map(TrustProvider::trustOptions)
+                        .map(TrustOptions::toString).orElse("-"))
+                .map(options -> " (TLS: " + options + ") ").orElse("");
+        var cipherSuitesAllowed = tlsToSummarize.map(t -> Optional.ofNullable(t.cipherSuites())
+                        .map(AllowDeny::allowed).orElse(Collections.emptyList()))
+                .map(allowedCiphers -> " (Allowed Ciphers: " + allowedCiphers + ")").orElse("");
+        var cipherSuitesDenied = tlsToSummarize.map(t -> Optional.ofNullable(t.cipherSuites())
+                        .map(AllowDeny::denied).orElse(Collections.emptySet()))
+                .map(deniedCiphers -> " (Denied Ciphers: " + deniedCiphers + ")").orElse("");
+        var protocolsAllowed = tlsToSummarize.map(t -> Optional.ofNullable(t.protocols())
+                        .map(AllowDeny::allowed).orElse(Collections.emptyList()))
+                .map(protocols -> " (Allowed Protocols: " + protocols + ")").orElse("");
+        var protocolsDenied = tlsToSummarize.map(t -> Optional.ofNullable(t.protocols())
+                        .map(AllowDeny::denied).orElse(Collections.emptySet()))
+                .map(protocols -> " (Denied Protocols: " + protocols + ")").orElse("");
+
+        return tls + cipherSuitesAllowed + cipherSuitesDenied + protocolsAllowed + protocolsDenied;
     }
 
     private static ClusterNetworkAddressConfigProvider buildAddressProviderService(@NonNull VirtualCluster virtualCluster,
